+++ src/search-strategies/looped-indexOf-anchored/index.ts
@@ -1,1 +1,2 @@
 export { LoopedIndexOfAnchoredSearchStrategy } from "./search-strategy.ts";
+export type { StringBufferState as StringAnchorSearchState } from "../string-buffer-strategy-base.ts";